requirements.txt
# 核心依赖 - Core Dependencies
pydantic>=2.0.0
jinja2>=3.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
tenacity>=8.4.0
sqlite3>=3.11.0

# 异步支持 - Async Support
asyncio-py>=3.11.0
aiofiles>=23.2.0
aiohttp>=3.9.0

# 可选扩展 - Optional Extensions
# 浏览器自动化 (如果需要)
# browser-use>=0.2.0

# API服务 (如果需要)
# fastapi>=0.104.0
# uvicorn>=0.24.0

# 开发工具 - Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 性能监控 - Performance Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# 类型检查 - Type Checking
types-requests>=2.31.0

# 日志和配置 - Logging and Configuration
pyyaml>=6.0
python-box>=7.0.0
click>=8.0.0
rich>=13.0.0